+++ app/components/layout/header-main.js
@@ -1,12 +1,14 @@
 import React from 'react';
 import { observer } from 'mobx-react/native';
-import { View, TouchableOpacity } from 'react-native';
+import { TouchableOpacity } from 'react-native';
 import Text from '../controls/custom-text';
 import SafeComponent from '../shared/safe-component';
+import CommonHeader from '../shared/common-header';
 import routerMain from '../routes/router-main';
 import { vars } from '../../styles/styles';
 import BackIcon from './back-icon';
 import testLabel from '../helpers/test-label';
+import icons from '../helpers/icons';
 
 @observer
 export default class HeaderMain extends SafeComponent {
@@ -16,17 +18,7 @@ export default class HeaderMain extends SafeComponent {
         if (routerMain.isBackVisible) {
             leftIcon = <BackIcon testID="buttonChatBack" />;
         }
-        const bigContainerStyle = {
-            paddingTop: vars.statusBarHeight,
-            height: vars.headerHeight,
-            flex: -1,
-            justifyContent: 'flex-end',
-            backgroundColor: vars.darkBlue
-        };
         const textStyle = {
-            flex: 1,
-            flexGrow: 1,
-            flexShrink: 1,
             color: vars.white,
             fontSize: vars.font.size.huge,
             textAlign: 'center',
@@ -35,55 +27,34 @@ export default class HeaderMain extends SafeComponent {
             borderColor: 'blue',
             borderWidth: 0,
             backgroundColor: 'transparent'
-            // marginLeft: routerMain.currentIndex === 0 ? vars.iconSize * 3 : -vars.iconSize
-        };
-        const containerStyle = {
-            flex: -1,
-            flexDirection: 'row',
-            alignItems: 'center',
-            paddingTop: vars.statusBarHeight,
-            minHeight: vars.headerHeight
         };
         const actionContainerStyle = {
-            flex: 1,
             flexGrow: 1,
-            flexShrink: 1,
             flexDirection: 'row',
-            alignItems: 'center'
+            alignItems: 'center',
+            justifyContent: 'center'
+        };
+        const outerStyle = {
+            backgroundColor: vars.darkBlue
         };
         const { titleAction } = routerMain;
-        const iconPlaceHolderWidth = (32 + vars.iconSize);
-        const paddingRight = !rightIcon && leftIcon ? iconPlaceHolderWidth : 0;
-        const paddingLeft = rightIcon && !leftIcon ? iconPlaceHolderWidth : 0;
         const title = this.props.title || routerMain.title;
-        return (
-            <View style={bigContainerStyle}>
-                <View key={`header_${routerMain.route}_${routerMain.currentIndex}`} style={containerStyle}>
-                    <View style={{ flex: 0 }}>
-                        {leftIcon}
-                    </View>
-                    <TouchableOpacity
-                        {...testLabel(title)}
-                        style={actionContainerStyle}
-                        activeOpacity={titleAction ? 0.2 : 1}
-                        onPress={titleAction}
-                        pressRetentionOffset={vars.retentionOffset}>
-                        <Text semibold
-                            ellipsizeMode="middle"
-                            numberOfLines={1}
-                            style={[textStyle, { paddingRight, paddingLeft }]}>
-                            {title}
-                            {titleAction && <Text style={{ fontSize: vars.font.size.normal }}>
-                                {'  ▼  '}
-                            </Text>}
-                        </Text>
-                    </TouchableOpacity>
-                    <View style={{ flex: 0 }}>
-                        {rightIcon}
-                    </View>
-                </View>
-            </View>
+        const titleComponent = (
+            <TouchableOpacity
+                {...testLabel(title)}
+                style={actionContainerStyle}
+                activeOpacity={titleAction ? 0.2 : 1}
+                onPress={titleAction}
+                pressRetentionOffset={vars.retentionOffset}>
+                <Text semibold ellipsizeMode="middle" numberOfLines={1} style={textStyle}>
+                    {title}
+                </Text>
+                {titleAction && icons.whiteNoPadding('arrow-drop-down', titleAction)}
+            </TouchableOpacity>
         );
+        // this is for animation purposes so that object gets completely redrawn on transition
+        const unique = `header_${routerMain.route}_${routerMain.currentIndex}`;
+        return <CommonHeader {...{ unique, titleComponent, leftIcon, rightIcon, outerStyle }} />;
     }
 }