+++ app/components/messaging/chat-item.js
@@ -8,15 +8,14 @@ import contactState from '../contacts/contact-state';
 import fileState from '../files/file-state';
 import { systemMessages } from '../../lib/icebear';
 import IdentityVerificationNotice from './identity-verification-notice';
-import FolderInlineContainer from '../files/folder-inline-container';
 import { vars } from '../../styles/styles';
 @observer
 export default class ChatItem extends SafeComponent {
     setRef = ref => { this._ref = ref; };
 
     renderThrow() {
-        const { chatId, message } = this.props;
-        if (!message || !chatId) return null;
+        const { chat, message } = this.props;
+        if (!message || !chat) return null;
         const i = this.props.message;
         if (!i.sender) return null;
         const key = i.id;
@@ -25,16 +24,13 @@ export default class ChatItem extends SafeComponent {
         const onPressAvatar = () => contactState.contactView(i.sender);
         const onPress = i.sendError ? this.props.onRetryCancel : null;
 
-        if (i.systemData && i.systemData.action === 'folder') {
-            return <FolderInlineContainer {...i.systemData} />;
-        }
         // this causes double update on add message
         const error = !!i.signatureError;
         const systemMessageText =
             i.systemData && systemMessages.getSystemMessageText(i) || null;
         const videoCallLink = i.systemData && i.systemData.link || null;
         const files = i.files && i.files.map(
-            id => fileState.store.getByIdInChat(id, chatId)
+            id => fileState.store.getByIdInChat(id, chat.id)
         ).filter(f => f) || [];
         const images = files.filter(f => f.isImage) || [];
         const normalFiles = files.filter(f => !f.isImage) || [];
@@ -43,7 +39,7 @@ export default class ChatItem extends SafeComponent {
             firstImage = i.externalImages[0];
         }
         const hasDeletedFile = i.files && !i.files.find(
-            id => fileState.store.getByIdInChat(id, chatId)
+            id => fileState.store.getByIdInChat(id, chat.id)
         );
         const shouldDisplayIdentityNotice = i.systemData && i.systemData.action === 'join';
 
@@ -56,6 +52,7 @@ export default class ChatItem extends SafeComponent {
                     contact={i.sender}
                     isDeleted={i.sender ? i.sender.isDeleted : false}
                     files={normalFiles.map(f => f.fileId)}
+                    folders={i.folders}
                     inlineImage={firstImage}
                     receipts={i.receipts}
                     hideOnline
@@ -65,7 +62,8 @@ export default class ChatItem extends SafeComponent {
                     message={text}
                     hasDeletedFile={hasDeletedFile}
                     isChat
-                    chatId={chatId}
+                    chat={chat}
+                    messageObject={i}
                     fullnameIsBold
                     systemMessage={systemMessageText}
                     videoCallMessage={videoCallLink}